+++ ground-control/app/main.py
@@ -39,7 +39,7 @@ CONFIG_PATHS = [
     "/share",
 ]
 
-VERSION = "0.1.10"
+VERSION = "0.1.11"
 
 # Global state
 state: TaskState = TaskState()